+++ dmcr/datamodels/evaluator/__init__.py
@@ -1,0 +1,2 @@
+from dmcr.datamodels.evaluator.BaseDatamodelsEvaluator import BaseDatamodelsEvaluator
+from dmcr.datamodels.evaluator.LinearRegressorEvaluator import LinearRegressorEvaluator
