+++ dmcr/datamodels/pipeline/__init__.py
@@ -1,0 +1,1 @@
+from dmcr.datamodels.pipeline.Datamodels import DatamodelPipeline
