+++ dmcr/evaluator/BaseEvaluator.py
