+++ dmcr/evaluator/__init__.py
@@ -1,0 +1,3 @@
+from dmcr.evaluator.BaseEvaluator import BaseEvaluator
+from dmcr.evaluator.Rouge_L_evaluator import Rouge_L_evaluator
+from dmcr.evaluator.GleuEvaluator import GleuEvaluator
