+++ dmcr/pipeline/BasePipeline.py
