+++ dmcr/pipeline/__init__.py
@@ -1,0 +1,3 @@
+from dmcr.pipeline.BasePipeline import BasePipeline
+from dmcr.pipeline.BaselinePipeline import BaselinePipeline
+from dmcr.pipeline.RAGPipeline import RAGPipeline
