+++ dmcr/retriever/__init__.py
@@ -1,0 +1,3 @@
+from dmcr.retriever.BaseRetriever import BaseRetriever
+from dmcr.retriever.BM25 import BM25
+from dmcr.retriever.NaiveDatamodelsRetriever import NaiveDatamodelsRetriever
