+++ dmcr/vector_store/__init__.py
@@ -1,0 +1,2 @@
+from dmcr.vector_store.BaseVectorStore import BaseVectorStore
+from dmcr.vector_store.Chroma import ChromaVectorStore
