--- src/datamodels/__init__.py
@@ -1,1 +1,0 @@
-from src.datamodels.config import DatamodelConfig
