--- src/datamodels/evaluator/__init__.py
@@ -1,2 +1,0 @@
-from src.datamodels.evaluator.BaseDatamodelsEvaluator import BaseDatamodelsEvaluator
-from src.datamodels.evaluator.LinearRegressorEvaluator import LinearRegressorEvaluator
