--- src/datamodels/models/__init__.py
@@ -1,1 +1,0 @@
-from src.datamodels.models.LinearRegressor import LinearRegressor
