--- src/datamodels/pipeline/__init__.py
@@ -1,1 +1,0 @@
-from src.datamodels.pipeline.Datamodels import DatamodelPipeline
