--- src/evaluator/__init__.py
@@ -1,3 +1,0 @@
-from src.evaluator.BaseEvaluator import BaseEvaluator
-from src.evaluator.Rouge_L_evaluator import Rouge_L_evaluator
-from src.evaluator.GleuEvaluator import GleuEvaluator
