--- src/llms/BaseLLM.py
@@ -1,8 +1,0 @@
-class BaseLLM:
-
-    def __init__(self) -> None:
-        pass
-        
-    
-    def run(self, input: str) -> str:
-        pass
