--- src/llms/GPT2Parser.py
@@ -1,45 +1,0 @@
-from src.llms import BaseLLM
-from transformers import AutoTokenizer, AutoModelForCausalLM
-from langchain_huggingface import HuggingFacePipeline
-from transformers import pipeline
-from transformers import BitsAndBytesConfig
-import torch
-from accelerate import Accelerator
-import os
-
-
-
-
-class GPT2Parser(BaseLLM):
-
-    def __init__(
-            self,
-            path = "../../models/llms/gpt2-large",
-        ) -> None:
-
-        self.tokenizer = AutoTokenizer.from_pretrained(path, max_length=1024, truncation=True)
-        self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
-        self.accelerator = Accelerator()
-        self.model = AutoModelForCausalLM.from_pretrained(
-                path, 
-                device_map={"": self.accelerator.process_index},
-                torch_dtype=torch.bfloat16,
-                use_safetensors=True,
-
-            )
-
-    
-    def run(self, input: str) -> str:
-
-
-        pipe = pipeline("text-generation",
-                        model = self.model,
-                        tokenizer = self.tokenizer,
-                        return_full_text=False,
-                        
-                )
-        input = f"Return a JSON object with an 'answer' key that answers straight forward the user input \n {input}"
-        tokenizer_kwargs = {'padding': True, 'truncation': True, 'max_length': 1024}
-        output = pipe( input, max_new_tokens=20, return_full_text=False, **tokenizer_kwargs)
-
-        return output[0]["generated_text"]
