--- src/retriever/__init__.py
@@ -1,3 +1,0 @@
-from src.retriever.BaseRetriever import BaseRetriever
-from src.retriever.BM25 import BM25
-from src.retriever.NaiveDatamodelsRetriever import NaiveDatamodelsRetriever
