--- src/vector_store/__init__.py
@@ -1,2 +1,0 @@
-from src.vector_store.BaseVectorStore import BaseVectorStore
-from src.vector_store.Chroma import ChromaVectorStore
